Add mxml v3.2

+++ phlib/mxml/mxml-index.c
@@ -330,8 +330,7 @@ mxmlIndexNew(mxml_node_t *node,		/* I - XML node tree */
 
   if ((ind = calloc(1, sizeof(mxml_index_t))) == NULL)
   {
-    mxml_error("Unable to allocate %d bytes for index - %s",
-               sizeof(mxml_index_t), strerror(errno));
+    mxml_error("Unable to allocate %d bytes for index - %s", (int)sizeof(mxml_index_t), strerror(errno));
     return (NULL);
   }
 
@@ -358,9 +357,7 @@ mxmlIndexNew(mxml_node_t *node,		/* I - XML node tree */
         * Unable to allocate memory for the index, so abort...
     */
 
-        mxml_error("Unable to allocate %d bytes for index: %s",
-               (ind->alloc_nodes + 64) * sizeof(mxml_node_t *),
-           strerror(errno));
+        mxml_error("Unable to allocate %d bytes for index: %s", (int)((ind->alloc_nodes + 64) * sizeof(mxml_node_t *)), strerror(errno));
 
         mxmlIndexDelete(ind);
     return (NULL);
